@@ -1,4 +1,4 @@
-class InitialSchema < ActiveRecord::Migration
+class InitialSchema < ActiveRecord::Migration[7.0]
   def change
     create_table :categories do |t|
       t.string :name, null: false
